refactor: 계산 결과 탭 내의 계산 결과 row와 추천 상품 목록 추상화
계산 결과 row 간의 간격도 Spacing으로 조정하게 될 가능성이 있어보이고, 3가지 결과 계산 로직을 하나의 컴포넌트에서 처리하기 보다는 각각의 컴포넌트에서 책임을 가지는게 좋을 것 같아서 row별로 추상화 진행.

추천 상품 목록 header와 리스트도 Spacing이 있기 때문에 header를 추상화 하지 않을까 고민했지만, 헤더는 리스트의 일부이기 때문에 리스트와 함께 두는게 응집도가 높다고 판단하여 함께 유지함.
