+++ src/pages/savingsCalculator/components/CalculationResultEmptyProduct.tsx
@@ -1,0 +1,5 @@
+import { ListRow } from 'tosslib';
+
+export function CalculationResultEmptyProduct() {
+  return <ListRow contents={<ListRow.Texts type="1RowTypeA" top="상품을 선택해주세요." />} />;
+}
